r11967: Fix more 64-bit warnings. (This used to be commit 9c4436a124f874ae240feaf590141d48c33a635f)

+++ source4/dsdb/samdb/samdb.c
@@ -800,7 +800,7 @@ int samdb_msg_add_uint(struct ldb_context *sam_ldb, TALLOC_CTX *mem_ctx, struct 
 int samdb_msg_add_int64(struct ldb_context *sam_ldb, TALLOC_CTX *mem_ctx, struct ldb_message *msg,
 			const char *attr_name, int64_t v)
 {
-	const char *s = talloc_asprintf(mem_ctx, "%lld", v);
+	const char *s = talloc_asprintf(mem_ctx, "%lld", (long long)v);
 	return samdb_msg_add_string(sam_ldb, mem_ctx, msg, attr_name, s);
 }
 
@@ -810,7 +810,7 @@ int samdb_msg_add_int64(struct ldb_context *sam_ldb, TALLOC_CTX *mem_ctx, struct
 int samdb_msg_add_uint64(struct ldb_context *sam_ldb, TALLOC_CTX *mem_ctx, struct ldb_message *msg,
 			const char *attr_name, uint64_t v)
 {
-	const char *s = talloc_asprintf(mem_ctx, "%llu", v);
+	const char *s = talloc_asprintf(mem_ctx, "%llu", (unsigned long long)v);
 	return samdb_msg_add_string(sam_ldb, mem_ctx, msg, attr_name, s);
 }
